--- src/assets/js/firebase.js
@@ -1,9 +1,0 @@
-const firebase = require("firebase/app");
-require("firebase/auth");
-const firebaseConfig = require('../../env/env.firebase');
-
-if (firebase.apps.length === 0) {
-  firebase.initializeApp(firebaseConfig);
-}
-
-module.exports = { firebase };
