+++ src/core/logger.js
@@ -1,11 +1,13 @@
 const JSZip = require('jszip');
-const Bowser = require("bowser");
+const Bowser = require('bowser');
 const events = JSON.parse(localStorage.getItem('CONFIG-EVENTS') || '[]');
 let log = '';
 
 exports.write = (...args) => {
   const date = `[${new Date().toISOString()}] `;
-  const msg = args.map(msg => typeof msg === 'object' ? JSON.stringify(msg) : msg).join(', ');
+  const msg = args
+    .map((msg) => (typeof msg === 'object' ? JSON.stringify(msg) : msg))
+    .join(', ');
   log += date + msg + '\n';
 };
 
@@ -14,7 +16,7 @@ exports.getLog = () => log;
 exports.saveLogs = () => {
   let eventKey = 'Setup';
   if (events.length > 0 && events[0].toa_event_key) {
-    eventKey = events[0].toa_event_key
+    eventKey = events[0].toa_event_key;
   }
 
   const fileName = `${eventKey} Logs.zip`;
@@ -23,7 +25,7 @@ exports.saveLogs = () => {
   zip.file('browser.json', JSON.stringify(getBowserDetails(), null, 2));
   zip.file('console.txt', log);
 
-  zip.generateAsync({type: 'blob'}).then((blob) => {
+  zip.generateAsync({ type: 'blob' }).then((blob) => {
     // Save the file with a new name
     const tag = document.createElement('a');
     tag.download = fileName;
@@ -40,6 +42,6 @@ function getBowserDetails() {
     ...browser,
     WebSocketSupport: !!window.WebSocket,
     ram: navigator.deviceMemory,
-    onLine: navigator.onLine
-  }
+    onLine: navigator.onLine,
+  };
 }
