+++ src/core/match-details.js
@@ -1,4 +1,4 @@
-const apis = require('../../apis');
+const apis = require('./apis');
 const events = JSON.parse(localStorage.getItem('CONFIG-EVENTS'));
 const toaApi = apis.toa;
 
@@ -15,7 +15,10 @@ module.exports = async function uploadMatchDetails(match, matchKey, eventKey) {
     data = getMatchDetails1920(match, matchKey);
   }
 
-  return toaApi(apiKeys[eventKey]).put(`/event/${eventKey}/matches/${matchKey}/details`, JSON.stringify([data]));
+  return toaApi(apiKeys[eventKey]).put(
+    `/event/${eventKey}/matches/${matchKey}/details`,
+    JSON.stringify([data])
+  );
 };
 
 function getMatchDetails1819(details, matchKey) {
@@ -23,14 +26,27 @@ function getMatchDetails1819(details, matchKey) {
     let partCrater = 0;
     let fullCrater = 0;
     switch (parking) {
-      case 15: partCrater = 1; break;
-      case 25: fullCrater = 1; break;
-      case 30: partCrater = 2; break;
-      case 40: partCrater = 1; fullCrater = 1; break;
-      case 50: fullCrater = 2; break;
-      default: partCrater = parking; break; // We should never get here
+      case 15:
+        partCrater = 1;
+        break;
+      case 25:
+        fullCrater = 1;
+        break;
+      case 30:
+        partCrater = 2;
+        break;
+      case 40:
+        partCrater = 1;
+        fullCrater = 1;
+        break;
+      case 50:
+        fullCrater = 2;
+        break;
+      default:
+        partCrater = parking;
+        break; // We should never get here
     }
-    return {partCrater, fullCrater};
+    return { partCrater, fullCrater };
   };
 
   return {
@@ -67,14 +83,22 @@ function getMatchDetails1819(details, matchKey) {
 
 function getMatchDetails1920(details, matchKey) {
   const getCapLevel = (alliance) => {
-    const level1 = alliance.robot1.capstoneLevel === -1 ? 0 : alliance.robot1.capstoneLevel;
-    const level2 = alliance.robot2.capstoneLevel === -1 ? 0 : alliance.robot2.capstoneLevel;
+    const level1 =
+      alliance.robot1.capstoneLevel === -1 ? 0 : alliance.robot1.capstoneLevel;
+    const level2 =
+      alliance.robot2.capstoneLevel === -1 ? 0 : alliance.robot2.capstoneLevel;
     return level1 + level2;
   };
   const getSkystones = (alliance) => {
-    if (alliance.autoStones[0] === 'SKYSTONE' && alliance.autoStones[1] === 'SKYSTONE') {
+    if (
+      alliance.autoStones[0] === 'SKYSTONE' &&
+      alliance.autoStones[1] === 'SKYSTONE'
+    ) {
       return 2;
-    } else if (alliance.autoStones[0] === 'SKYSTONE' || alliance.autoStones[1] === 'SKYSTONE') {
+    } else if (
+      alliance.autoStones[0] === 'SKYSTONE' ||
+      alliance.autoStones[1] === 'SKYSTONE'
+    ) {
       return 1;
     } else {
       return 0;
@@ -82,7 +106,8 @@ function getMatchDetails1920(details, matchKey) {
   };
   const getStones = (alliance) => {
     const skystones = getSkystones(alliance);
-    const stones = alliance.autoStones.filter(x => x && x !== 'NONE').length - skystones;
+    const stones =
+      alliance.autoStones.filter((x) => x && x !== 'NONE').length - skystones;
     return Math.max(stones, 0); // Prevent negative numbers
   };
 
@@ -114,12 +139,12 @@ function getMatchDetails1920(details, matchKey) {
       robot_1: {
         nav: details.red.robot1.navigated,
         parked: details.red.robot1.parked,
-        cap_level: details.red.robot1.capstoneLevel
+        cap_level: details.red.robot1.capstoneLevel,
       },
       robot_2: {
         nav: details.red.robot2.navigated,
         parked: details.red.robot2.parked,
-        cap_level: details.red.robot2.capstoneLevel
+        cap_level: details.red.robot2.capstoneLevel,
       },
       foundation_moved: details.red.foundationMoved,
       auto_transport_points: details.red.autoTransportPoints,
@@ -129,14 +154,15 @@ function getMatchDetails1920(details, matchKey) {
       tele_transport_points: details.red.driverControlledTransportPoints,
       tele_placed_points: details.red.driverControlledPlacedPoints,
       tower_bonus: details.red.towerBonusPoints / 2,
-      tower_capping_bonus: (details.red.capstonePoints - getCapLevel(details.red)) / 5,
+      tower_capping_bonus:
+        (details.red.capstonePoints - getCapLevel(details.red)) / 5,
       tower_level_bonus: getCapLevel(details.red),
       end_robots_parked: details.red.parkingPoints / 5,
       auto_points: details.red.autonomousPoints,
       auto_total: details.red.auto,
       tele_total: details.red.tele,
       end_total: details.red.end,
-      penalty_total: details.red.penalty
+      penalty_total: details.red.penalty,
     },
 
     blue: {
@@ -158,12 +184,12 @@ function getMatchDetails1920(details, matchKey) {
       robot_1: {
         nav: details.blue.robot1.navigated,
         parked: details.blue.robot1.parked,
-        cap_level: details.blue.robot1.capstoneLevel
+        cap_level: details.blue.robot1.capstoneLevel,
       },
       robot_2: {
         nav: details.blue.robot2.navigated,
         parked: details.blue.robot2.parked,
-        cap_level: details.blue.robot2.capstoneLevel
+        cap_level: details.blue.robot2.capstoneLevel,
       },
       foundation_moved: details.blue.foundationMoved,
       auto_transport_points: details.blue.autoTransportPoints,
@@ -173,15 +199,16 @@ function getMatchDetails1920(details, matchKey) {
       tele_transport_points: details.blue.driverControlledTransportPoints,
       tele_placed_points: details.blue.driverControlledPlacedPoints,
       tower_bonus: details.blue.towerBonusPoints / 2,
-      tower_capping_bonus: (details.blue.capstonePoints - getCapLevel(details.blue)) / 5,
+      tower_capping_bonus:
+        (details.blue.capstonePoints - getCapLevel(details.blue)) / 5,
       tower_level_bonus: getCapLevel(details.blue),
       end_robots_parked: details.blue.parkingPoints / 5,
       auto_points: details.blue.autonomousPoints,
       auto_total: details.blue.auto,
       tele_total: details.blue.tele,
       end_total: details.blue.end,
-      penalty_total: details.blue.penalty
+      penalty_total: details.blue.penalty,
     },
-    randomization: details.randomization
+    randomization: details.randomization,
   };
 }
