+++ src/index.html
@@ -8,9 +8,9 @@
   <meta property="og:description" content="Sync the your event with The Orange Alliance in realtime."/>
   <title>DataSync</title>
   <link rel="shortcut icon" href="assets/imgs/favicon.png">
-  <link href="assets/css/styles.css" rel="stylesheet">
-  <link href="assets/css/bootstrap.min.css" rel="stylesheet">
-  <link href="assets/material-components-web@2.2.0/material-components-web.min.css" rel="stylesheet">
+  <link href="/assets/css/styles.css" rel="stylesheet">
+  <link href="/assets/css/bootstrap.min.css" rel="stylesheet">
+  <link href="/assets/material-components-web@2.2.0/material-components-web.min.css" rel="stylesheet">
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet" >
 </head>
 <body>
@@ -21,7 +21,7 @@
   </div>
 </div>
 
-<script src="assets/material-components-web@2.2.0/material-components-web.min.js"></script>
+<script src="/assets/material-components-web@2.2.0/material-components-web.min.js"></script>
 
 </body>
 </html>
