+++ src/setup.html
@@ -8,10 +8,10 @@
   <meta property="og:description" content="Sync the your event with The Orange Alliance in realtime."/>
   <title>DataSync</title>
   <link rel="shortcut icon" href="../assets/imgs/favicon.png">
-  <link href="assets/css/styles.css" rel="stylesheet">
-  <link href="assets/mdi/css/materialdesignicons.min.css" rel="stylesheet">
-  <link href="assets/css/bootstrap.min.css" rel="stylesheet">
-  <link href="assets/material-components-web@2.2.0/material-components-web.min.css" rel="stylesheet">
+  <link href="/assets/css/styles.css" rel="stylesheet">
+  <link href="/assets/mdi/css/materialdesignicons.min.css" rel="stylesheet">
+  <link href="/assets/css/bootstrap.min.css" rel="stylesheet">
+  <link href="/assets/material-components-web@2.2.0/material-components-web.min.css" rel="stylesheet">
 </head>
 <body class="setup-page">
 <div class="main-content row">
@@ -105,6 +105,6 @@
   </div>
 </div>
 
-<script src="../assets/material-components-web@2.2.0/material-components-web.min.js"></script>
+<script src="/assets/material-components-web@2.2.0/material-components-web.min.js"></script>
 </body>
 </html>
