+++ src/sync.html
@@ -5,12 +5,12 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <title>DataSync</title>
   <link rel="shortcut icon" href="assets/imgs/favicon.png">
-  <link href="assets/css/sync-page.css" rel="stylesheet">
-  <link href="assets/css/styles.css" rel="stylesheet">
-  <link href="assets/css/settings.css" rel="stylesheet">
-  <link href="assets/mdi/css/materialdesignicons.min.css" rel="stylesheet">
-  <link href="assets/css/bootstrap.min.css" rel="stylesheet">
-  <link href="assets/material-components-web@2.2.0/material-components-web.min.css" rel="stylesheet">
+  <link href="/assets/css/sync-page.css" rel="stylesheet">
+  <link href="/assets/css/styles.css" rel="stylesheet">
+  <link href="/assets/css/settings.css" rel="stylesheet">
+  <link href="/assets/mdi/css/materialdesignicons.min.css" rel="stylesheet">
+  <link href="/assets/css/bootstrap.min.css" rel="stylesheet">
+  <link href="/assets/material-components-web@2.2.0/material-components-web.min.css" rel="stylesheet">
 </head>
 <body class="sync-page">
 <header class="mdc-top-app-bar" id="status-header">
@@ -227,6 +227,6 @@
   </div>
 </div>
 
-<script src="assets/material-components-web@2.2.0/material-components-web.min.js"></script>
+<script src="/assets/material-components-web@2.2.0/material-components-web.min.js"></script>
 </body>
 </html>
